New main menu implementation.
